Created separate library project and test project
