+++ C#/TollCalculator/ITollCalculator.cs
@@ -1,0 +1,9 @@
+using System;
+namespace TollFeeCalculator
+{
+    public interface ITollCalculator
+    {
+        public int GetTollFee(IVehicle vehicle, DateTime[] dates);
+        public int GetTollFee(DateTime date, IVehicle vehicle);
+    }
+}
