+++ C#/TollCalculator/IVehicle.cs
